backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pandas==2.1.4
python-dotenv==1.0.0
google-generativeai==0.3.2
werkzeug==3.0.1
bcrypt==4.1.2
pymongo==4.6.1
